moved AR components to AnnualReport/
