+++ coops-ui/src/components/AnnualReport/ARComplete.vue
