+++ coops-ui/src/components/AnnualReport/RegisteredOfficeAddress.vue
